third_party/WebKit/LayoutTests/media/track/text-track-cue-constructor.html

<!DOCTYPE html>
<html>
    <head>
        <title>TextTrackCue constructor</title>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <div id="log"></div>
        <script>
            test(function()
            {
                assert_not_equals(TextTrackCue, VTTCue);
            }, "TextTrackCue and VTTCue are separate interfaces");
            test(function()
            {
	        var cue = new TextTrackCue(0, 0, "");
                assert_true(cue instanceof TextTrackCue);
                assert_true(cue instanceof VTTCue);
            }, "TextTrackCue constructor returns a VTTCue");
            test(function()
            {
                try { 
                    new TextTrackCue(0);
                    assert_unreached("Expected a TypeError");
                } catch (e) {
                    assert_true(e instanceof TypeError);
                    assert_equals(e.message, "Failed to construct 'TextTrackCue': 3 arguments required, but only 1 present.");
                }
            }, "When not fully applied, TextTrackCue constructor throws a TypeError");
        </script>
    </body>
</html>
